docs: extract local vite plugins from `vite.config.ts`
